requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
pygments==2.17.2
rich==13.7.1
shellingham==1.5.4
typer==0.9.0
regex==2023.12.25
prompt-toolkit==3.0.43
wcwidth==0.2.13
python-dotenv==1.0.1
python-crontab==3.0.0
InquirerPy==0.3.4
ruff==0.3.2
protobuf~=5.28.3
requests==2.31.0
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.151.0